images/h2pc-agent-llm/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0.0,<3.0.0"]

[tool.poetry]
name = "h2pc-agent-llm"
authors = ["Ho Kim <ho.kim@ulagbulag.io>"]
description = "H2PC LLM Agent"
homepage = "https://github.com/SmartX-Team/smartx-k8s"
keywords = ["h2pc", "kubernetes", "pipeline"]
license = "GPL-3.0-or-later WITH Classpath-exception-2.0"
readme = "README.md"
repository = "https://github.com/SmartX-Team/smartx-k8s"
version = "2.0.0-alpha.1"

packages = [{ include = "main.py" }]

[tool.poetry.dependencies]
# Compatible Python versions
python = ">=3.12,<4.0"

crc32c = { version = ">=2.7.1,<3.0.0" }
deltalake = { version = ">=0.25.5,<0.26.0" }
kafka-python = { version = ">=2.1.5,<3.0.0" }
ollama = { version = ">=0.4.8,<0.5.0" }
openai = { version = ">=1.75.0,<2.0.0" }
pydantic = { version = ">=2.11.3,<3.0.0" }
pydantic-settings = { version = ">=2.8.1,<3.0.0" }
pyiceberg = { version = ">=0.9.0,<0.10.0" }
pyyaml = { version = ">=6.0.2,<6.1.0" }

[tool.poetry.group.dev.dependencies]
python-dotenv = ">=1.1.0,<1.2.0"
